GAL-833# 4 unit tests successfully passing
