+++ package.json
@@ -16,6 +16,7 @@
     "@aws-sdk/client-api-gateway": "^3.485.0",
     "@aws-sdk/client-dynamodb": "^3.485.0",
     "@aws-sdk/client-lambda": "^3.484.0",
+    "@aws-sdk/client-secrets-manager": "^3.495.0",
     "@aws-sdk/config-resolver": "^3.374.0",
     "@aws-sdk/util-stream-node": "^3.374.0",
     "@babel/plugin-syntax-import-assertions": "^7.23.3",
